Fix race in conformance/interfaces/pthread_mutex_init/1-2.c
This test was running deadlk_issue several times and expecting the
returned global to be zero if `pthread_mutex_lock` has not returned,
however the resetting on `returned` to zero was happening only after the
`sem_post`.  The main thread is calling `sched_yeild` which on most
platforms is enough to cover up this issue, but not with WebAssembly.

Some alternatives to landing this change:

1. We could just disable this test.
2. We could make sched_yield actualy do something rather than be
   a no-op.